Try Sentence transformers + some refactoring.
- Sentence transformer works better than biobert last 4 layers average embeddings
- and is ~ 3 times faster, probably due to batching
- Still it seems to cut the input size to 512 tokens
